refactor: 완료페이지 Detail 데이터 수정
